revos-server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai>=1.50.0
pinecone==5.3.1
flask-sqlalchemy==3.0.5
PyJWT>=2.8.1
PyPDF2==3.0.1
python-docx==1.1.2
pdfplumber==0.11.4
plaid-python==37.0.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.100.0
icalendar==5.0.7
pytz==2024.1
